% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koopman.R
\name{Koopman_MM1}
\alias{Koopman_MM1}
\title{Koopman formulas as described in Koopman, 1971}
\usage{
Koopman_MM1(lambda, mu, C, breaks, init_prob = c(1, rep(0, C - 1)))
}
\arguments{
\item{lambda}{rate of arrivals}

\item{mu}{rate of service}

\item{C}{capacity of system}

\item{breaks}{number of times at which ode is calculated}

\item{init_prob}{vector of initial probabilities}
}
\description{
Koopman formulas as described in Koopman, 1971
}
\examples{
z <- Koopman_MM1(1.5, 2, 20, 20)
colSums(z)
}
